move class QueryImpl into separate file
